add post process effects and other stuff
