+++ Assets/Materials/Lightbulb.mat
@@ -77,5 +77,5 @@ Material:
     - _ZWrite: 1
     m_Colors:
     - _Color: {r: 1, g: 1, b: 1, a: 1}
-    - _EmissionColor: {r: 0.7830189, g: 0.7830189, b: 0.7830189, a: 1}
+    - _EmissionColor: {r: 0.7749607, g: 0.7749607, b: 0.7749607, a: 1}
   m_BuildTextureStacks: []
