+++ Assets/Models/_Done/FrontCounter/FrontCounterMenu.mat
@@ -109,5 +109,5 @@ Material:
     - _isMetallic: 0
     m_Colors:
     - _Color: {r: 1, g: 1, b: 1, a: 1}
-    - _EmissionColor: {r: 0.9056604, g: 0.9002728, b: 0.80740476, a: 1}
+    - _EmissionColor: {r: 0.42426407, g: 0.42149112, b: 0.37712362, a: 1}
   m_BuildTextureStacks: []
