+++ Assets/Models/_Structures/Floor/Floor.mat
@@ -106,12 +106,12 @@ Material:
     - _OcclusionStrength: 1
     - _Parallax: 0.02
     - _Progress_KEEP_ZERO: 0
-    - _RoughnessMultiplier: 0.3
+    - _RoughnessMultiplier: 0.2
     - _SmoothnessTextureChannel: 0
     - _SpecularHighlights: 1
     - _SrcBlend: 1
-    - _TransitionScaleX: 0.1
-    - _TransitionScaleY: 0.1
+    - _TransitionScaleX: 0.2
+    - _TransitionScaleY: 0.2
     - _UVSec: 0
     - _ZWrite: 1
     - _isMetallic: 0
