+++ Assets/Models/_Structures/Walls/WallInt.mat
@@ -93,7 +93,7 @@ Material:
     - _OcclusionStrength: 1
     - _Parallax: 0.02
     - _Progress_KEEP_ZERO: 0
-    - _RoughnessMultiplier: 0.2
+    - _RoughnessMultiplier: 0.25
     - _SmoothnessTextureChannel: 0
     - _SpecularHighlights: 1
     - _SrcBlend: 1
