+++ Assets/PostProcessProfiles/FutureProfile.asset
@@ -21,16 +21,16 @@ MonoBehaviour:
     value: 0
   color:
     overrideState: 1
-    value: {r: 0.49669412, g: 0, b: 0.754717, a: 1}
+    value: {r: 0.27540612, g: 0.13336597, b: 0.3490566, a: 1}
   center:
     overrideState: 0
     value: {x: 0.5, y: 0.5}
   intensity:
     overrideState: 1
-    value: 0.08
+    value: 0.459
   smoothness:
     overrideState: 1
-    value: 1
+    value: 0.825
   roundness:
     overrideState: 0
     value: 1
@@ -44,7 +44,7 @@ MonoBehaviour:
   opacity:
     overrideState: 0
     value: 1
---- !u!114 &-4283592780595630688
+--- !u!114 &-4609391066552857843
 MonoBehaviour:
   m_ObjectHideFlags: 3
   m_CorrespondingSourceObject: {fileID: 0}
@@ -97,10 +97,10 @@ MonoBehaviour:
     value: 1
   temperature:
     overrideState: 1
-    value: -10.5
+    value: -26
   tint:
     overrideState: 1
-    value: 15.8
+    value: 15.6
   colorFilter:
     overrideState: 0
     value: {r: 1, g: 1, b: 1, a: 1}
@@ -115,10 +115,10 @@ MonoBehaviour:
     value: 0
   postExposure:
     overrideState: 1
-    value: -1.3
+    value: -1.2
   contrast:
     overrideState: 1
-    value: -4.9
+    value: -15
   mixerRedOutRedIn:
     overrideState: 0
     value: 100
@@ -1369,4 +1369,31 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   settings:
   - {fileID: -5863607540590732168}
-  - {fileID: -4283592780595630688}
+  - {fileID: -4609391066552857843}
+  - {fileID: 5182158339184684570}
+--- !u!114 &5182158339184684570
+MonoBehaviour:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6050e2d5de785ce4d931e4dbdbf2d755, type: 3}
+  m_Name: ChromaticAberration
+  m_EditorClassIdentifier: 
+  active: 1
+  enabled:
+    overrideState: 1
+    value: 1
+  spectralLut:
+    overrideState: 0
+    value: {fileID: 0}
+    defaultState: 1
+  intensity:
+    overrideState: 1
+    value: 0.601
+  fastMode:
+    overrideState: 0
+    value: 0
