+++ Assets/Scenes/Level1/Lightmap-0_comp_dir.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1e2544312e81e779596235b133edb0b3a0b71445b4dfd6de886b47e97fb82537
-size 534391
+oid sha256:6cc2c1ef7829d1e24effcb8c4e0f588b83f89dd06ae5fc8bd4c5c0350f0d5162
+size 524667
