+++ Assets/Scenes/Level1/Lightmap-0_comp_light.exr
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9512d908799c5aa69faf0c91acc46a0d0e9d8682d9dcfefbb95be9625bd75a8e
-size 2071819
+oid sha256:6b15f20a7d6fdf788c70f4febc7517ce2a9db57a1b747cb52d1ca7a0538a28f3
+size 2040372
