+++ Assets/Scenes/Level1/Lightmap-1_comp_dir.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:fc7e7da3f5fed2e8f684f3cb2a3a3e3bf582f4de5e9f8ab25a421889c6c577fb
-size 647713
+oid sha256:88b72abded947903619e71d7886f153333438ee8ec3dbe9f5048237a5fce8a52
+size 620852
