+++ Assets/Scenes/Level1/Lightmap-1_comp_light.exr
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:7008a054220a50443f6be54fbe368add61c49ed36a1574f06df1ed3871ffc4b6
-size 2155059
+oid sha256:16e0dbda10508484bc37d78d591a08d403a6e6bb21eb5fbf39d9dbc6932727e4
+size 2069974
