+++ Assets/Scenes/Level1/ReflectionProbe-0.exr
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:cb5152ac341748ee30e69e91c160bd4b97854d0eff8e53391b54109879048b4e
-size 338977
+oid sha256:5d5f56be03a05b320c57e9bf0635c6f15b70fe9158e3eef9fc9f4ca3f38d7299
+size 338055
