+++ Assets/Scripts/CameraShader.cs
@@ -11,21 +11,14 @@ public class CameraShader : MonoBehaviour {
     [SerializeField] [MustBeAssigned] private Camera cam;
 
     [SerializeField] [ReadOnly] private bool isEffectEnabled = false;
-    private float initialFOV;
-    private float deltaFOV = 8;
-    private float finalFOV;
     private Coroutine currentShaderCoroutine;
-    private Coroutine currentFOVCoroutine;
     private Coroutine currentVolumeCoroutine;
 
     // Constants
     private const float SHADER_TRANSITION_MULTIPLIER = 1f;
     private const float VOLUME_TRANSITION_MULTIPLIER = 0.5f;
-    private const float FOV_TRANSITION_MULTIPLIER = 0.25f;
 
     void Start() {
-        initialFOV = cam.fieldOfView;
-        finalFOV = initialFOV + deltaFOV;
         cam.cullingMask = cam.cullingMask & ~(1 << LayerMask.NameToLayer("FPS")); // don't render the FPS layer
     }
 
@@ -40,9 +33,7 @@ public class CameraShader : MonoBehaviour {
         }
 
         if (currentShaderCoroutine != null) StopCoroutine(currentShaderCoroutine);
-        if (currentVolumeCoroutine != null) StopCoroutine(currentShaderCoroutine);
-        if (currentFOVCoroutine != null) StopCoroutine(currentShaderCoroutine);
-
+        if (currentVolumeCoroutine != null) StopCoroutine(currentVolumeCoroutine);
         OnSetEffectActive(enabled, transitionTime, onComplete);
 
         isEffectEnabled = enabled;
@@ -50,10 +41,8 @@ public class CameraShader : MonoBehaviour {
 
     private void OnSetEffectActive(bool active, float transitionTime, Action onComplete = null) {
         float end = active ? 1f : 0f;
-        float fovEnd = active ? finalFOV : initialFOV;
         currentShaderCoroutine = StartCoroutine(TransitionShader(end, transitionTime * SHADER_TRANSITION_MULTIPLIER, onComplete));
         currentVolumeCoroutine = StartCoroutine(TransitionVolume(end, transitionTime * VOLUME_TRANSITION_MULTIPLIER));
-        currentFOVCoroutine = StartCoroutine(TransitionFOV(fovEnd, transitionTime * FOV_TRANSITION_MULTIPLIER));
     }
 
     IEnumerator TransitionShader(float end, float duration, Action onComplete = null) {
@@ -72,22 +61,6 @@ public class CameraShader : MonoBehaviour {
         Shader.SetGlobalFloat("_Progress", end);
 
         onComplete?.Invoke();
-    }
-
-    IEnumerator TransitionFOV(float end, float duration) {
-        float elapsed_time = 0;
-        float currentProgress = cam.fieldOfView;
-
-        while (elapsed_time <= duration) {
-            cam.fieldOfView = Mathf.Lerp(currentProgress, end, elapsed_time / duration);
-
-            yield
-            return null; //Waits/skips one frame
-
-            elapsed_time += Time.deltaTime;
-        }
-
-        cam.fieldOfView = end;
     }
 
     IEnumerator TransitionVolume(float end, float duration) {
