+++ Assets/Scripts/CameraShaderFOVChange.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 67311d377a1670343ae28b84a680abe1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
