+++ ProjectSettings/TagManager.asset
@@ -20,10 +20,10 @@ TagManager:
   - Water
   - UI
   - Character
-  - WithFuturePostProcess
-  - NoFuturePostProcess
-  - Ground
+  - FuturePostProcess
   - 
+  - Ground
+  - DefaultPostProcess
   - 
   - 
   -
